service/llm_voice_agents/requirements.txt
videosdk-agents[deepgram,openai,elevenlabs,silero,turn_detector]
pydantic-settings
python-dotenv
PyJWT
videosdk-plugins-google
